Move GoGoDuck UserLog to own class
