+++ src/extension/gogoduck/src/main/java/au/org/emii/gogoduck/worker/GoGoDuck.java
@@ -420,45 +420,4 @@ public class GoGoDuck {
 
         return process.exitValue();
     }
-
-    public class UserLog {
-        private FileWriter userLogFileWriter;
-        UserLog() {
-            this.userLogFileWriter = null;
-        }
-
-        UserLog(String logFile) {
-            try {
-                this.userLogFileWriter = new FileWriter(logFile);
-            }
-            catch (IOException e) {
-                this.userLogFileWriter = null;
-                logger.info(String.format("Exception while opening user log at '%s', continuing anyway", logFile));
-            }
-        }
-
-        public void close() {
-            try {
-                if (null != userLogFileWriter) {
-                    userLogFileWriter.close();
-
-                }
-            }
-            catch (IOException e) {
-                logger.debug("Could not close user log file");
-            }
-        }
-
-        public void log(String s) {
-            if (null != userLogFileWriter) {
-                try {
-                    userLogFileWriter.write(s);
-                    userLogFileWriter.write(System.lineSeparator());
-                }
-                catch (IOException e) {
-                    logger.debug("Exception while writing to user log");
-                }
-            }
-        }
-    }
 }
